Updated openstack/openstack Project: openstack/oslo.serialization  5b0827a3df584b3e26bb44e5358cc28ac4a91636
jsonutils: replace strtime() usage with isoformat()

Following discussion in change I8b5119e64369ccac3423dccc04421f99912df733
in oslo.utils, we plan to deprecated and remove strtime() which is to be
considered dangerous.

This changes the json serialization code to embed the timezone in the
encoded datetime string if it's available.

Change-Id: I262391987d25b483d2bec4ac3abe94b6e9d032ad