Updated openstack/openstack Project: openstack/oslo.db  3cf218d0961cf1e0dc2c6f58602c307f43ac6ad9
Implement generic update-on-match feature

This feature provides the query.update_on_match()
and query.update_returning_pk() methods, as well
as the manufacture_persistent_object(),
manufacture_entity_criteria(), and manufacture_criteria()
utility functions.

query.update_on_match() is used to UPDATE a row based on a variety
of criteria, and to then return a fully persistent object
state representing the row that was matched.  It
essentially intends to provide an UPDATE that is guaranteed
to have matched a specific row in the presence of potential
race conditions without using any locking, and to then
return a record of that row as if it had been SELECTed.

query.update_returning_pk() is a public method that
also serves as part of the implementation of
query.update_on_match(); this method delivers an UPDATE
statement such that the primary key of the single row
matched is returned; if zero or multiple rows are matched,
and error is raised.   To handle this, several backend-specific
strategies are provided, which are automatically selected
based on the best available.  The lowest strategy
performs a re-SELECT, but still assumes there's
a simple unique column to be queried on, as is currently the
use case in Nova (uuid is present).  On Postgresql, MySQL
and other databases besides SQLite and possibly DB2, more
atomic strategies are used.

Change-Id: I059f4ae6e72cfa6681a179314144214639f283ef