Updated openstack/openstack Project: openstack/heat-translator  d064f39c45480d876d2acbca0b9b73f33485d366
TOSCA: Add missing artifacts references

The TOSCA wordpress template is missing artifacts. These artifacts were tested
with actual deployment but never added here. Unit tests are there for this
template for TOSCA parser side and we are working on to create tests for the
full translated template with HOT mapping.

Change-Id: I9ddedc6ee6e072e765d3d3fe6b97a2a7d8aa91d8