Updated openstack/openstack Project: openstack/cinder  5c394b148f43e655a2cb259f7c7d075e2b438362
Tests: Fix az test failure when PYTHONHASHSEED != 0

The test_list_availability_zones_enabled_service test calls
a method which returns a tuple that is generated from a dict.
This is compared to a tuple in the test, which can fail when
PYTHONHASHSEED is not 0, due to the tuples containing elements
in different orders.  (Because the dict used to build the tuple
was ordered differently.)

For this test, we can convert to lists and sort them to do the
comparison, since the ordering is not important.

Partial-Bug: #1348818

Change-Id: Ib7e5d59073e5e243450e28492d68942f6088bb5f