Updated openstack/openstack Project: openstack-infra/project-config  5e485c25d154efb1a5c3cc9028bc884841faba6e
zuul: add check-grenade-dsvm-neutron to tempest checks

Since check-grenade-dsvm-neutron is part of Neutron checks, we should
add this job to Tempest too because Javelin could break Grenade/Neutron
jobs.
A recent failure in Neutron gate was not detected, with this job we
would have been able to detect the issue and the patch would not have
been merged in Tempest.

Change-Id: Ic095fd1f42acb3748a7b7bf61d188155060d0d61