Updated openstack/openstack Project: openstack/barbican  78f90d8b204fcb56cfe19759a665044381cdd89c
Remove str() casting for the client_message variable

This is done since the e.client_message variable is a unicode
string that might be derived from a non-ascii input (for
example, the user-supplied paylaod content type). If the str()
function is applied to a non-ascii char, a UnicodeEncodeError
exception will be thrown.

Change-Id: I3ce5454a37347297b3ec0b0124234025a4aed872
Closes-bug: #1435613
Related-Bug: #1421478