Updated openstack/openstack Project: openstack/nova  385c8102d15e84b84dfd165ae1a2af03ffd5600a
Fix for deletes first preexisting port if second was attached to instance

In Ia5367cf064d40690670ffeac3c1f16998464c234, we added a capability
to preserve preexisting ports on server delete. There was one scenario
that broken as a result as reported in the bug.
"If I boot instance with preexisting port and attach one more port
then first port is deleted while I delete the instance."

Essentially we have to lookup get_preexisting_port_ids during
boot as well as when we are attaching an interface.

Closes-Bug: #1433524
Change-Id: I3d557dd95f442106c495249a5ef1d2ac36e6a2ea