Updated openstack/openstack Project: openstack/neutron  d18056de242eebefc1473fc41cd4c5047b215dce
Add no_delete flag to UT router context manager

Adds a no_delete flag to the router context manager in
the unit tests to allow specific tests to avoid executing
all of the cleanup logic for floating IPs and interfaces
just to exit the test.

Ultimately, this should become the default once we are
comfortable that there is adequate explicit unit tests
for the deletion logic.

Change-Id: Iecbcc77b67b440d726fa6e703a87884e0b0df321