Updated openstack/openstack Project: openstack/neutron  2e735da0ca1ce8783b89b2eb6157dd9a7df9c386
Fixes floating IP regression with multiple routers

During the refactor here:
Change-Id: I09e8a694cdff7f64a642a39b45cbd12422132806
Too much code was removed and caused floating ips to get miss assigned when
multiple routers with external networks in the same tenant are present. The
first router in the tenant was always being chosen. This patch adds back
some of the original code as well as a unit test.

Change-Id: I6f663cb1ce3e4a1340c415d13787a9855c4dcac2
Closes-Bug: 1422476