Updated openstack/openstack Project: openstack/manila  7eafabd6bb5bfc602cd5e7420f39a5b73f734fe8
Add share state verification for API 'unmanage'

Shares with transitional states like 'creating' or 'deleting' should not be
allowed to be unmanaged.
For the moment "manage/unmanage" operation is allowed only for driver mode
without handling of share servers. So, API operations against shares created
on top of share servers should be restricted on API level.

Change-Id: I73b554448eadcb96ef00f44535014e14dea91472
Closes-Bug: #1434642
Closes-Bug: #1434511