Updated openstack/openstack Project: openstack/ceilometer  cf51c4b98665f87a52864b082300e30cdfeee27e
exclude precise metaquery in query field

We use metaquery as a parameter of db func, and valid keys include
db func parameters.

When we parse query filters, the metadata dict is stored as key named
metaquery in kwargs, and we store q.field=foo as key named foo in kwargs.
Then if q.field=metaquery, the kwargs will get the key metaquery but
its type is something, for i.e. string, but not dict. However the db
func expects metaquery to be a dict but gets a string, then 500 is returned

This patch returns a 400 error for such case, and removes 'metaquery'
from valid keys in error message to avoid misleading end user.

Change-Id: I430717e71a6586d63f267baee29bb71ef8c9eac1
Closes-Bug: #1426466