Updated openstack/openstack Project: openstack/ceilometer  27ca273c049eb253fff603f15f54da78c0fb7abb
refuse to post sample which is not supported

Currently, if we remove wildcard * in pipeline.yaml and explictly
configure which metrics are supported, then post an unsupported sample
via API, then API will return 201 but that sample will not be stored
to database since pipeline don't support it.

This patch adds a check for each posted sample, if any of them is not
supported, then a 409 code will be returned.

Change-Id: I1f115929508c1ea432829d7d7dc2fe758e7d77ac
Closes-Bug: #1432077
DocImpact