Updated openstack/openstack Project: openstack/swift  66fb207fcb288889d73d42960b037e3ef7ef3ce2
Add lowest option to swift-recon disk usage output

Currently there is a "--top" option when running swift-recon for
disk usage stats. This option lists the x disks with the highest
disk usage in descending order.

This feature adds a "--lowest" option which does the opposite and
lists the y disks with lowest disk usage in ascending order.

Have also updated the docs section with --top and --lowest options

Change-Id: Ic15d407fe010a31995c2bdd9fb88548a1057f569