Updated openstack/openstack Project: openstack/governance  c61c7036d5dc81f932f4a7951184b823d084bed8
Add os-testr repo to QA

This commit adds a new repo under QA, os-testr. The intent of this repo
is to provide tooling to interact and use testr. Right now it contains
2 pieces subunit-trace, which has become the de-facto output filter
for running tests, and ostestr which is a testr wrapper similar to the
pretty_tox.sh scripts which are currently used everywhere but expands
on that functionality.

Change-Id: Idbd6251546809f6ea7a651be9e758619fddd4e7f