Updated openstack/openstack Project: openstack-infra/system-config  e82e1c511d3ef6af211964bc0cda7a391d8e07be
Cleanup devstack-(trusty|precise)-dib images

Right now we are using dib in hpcloud for devstack-trusty and
devstack-precise, this means we don't need the devstack-trusty-dib and
devstack-precise-dib labels and images to be provided by hpcloud. Remove
as much as we can while retaining the label without providers so that it
is there when we need to test dib images in rackspace.

Depends-On: I91f8d91318752e09a15d64c8b3c506cf83badca2
Change-Id: I7b82fb5162af908a86ffcb291d908a88867f3b06