Updated openstack/openstack Project: openstack/oslo-incubator  f0112390e8d3bdca2fd3a1d09f000b0bb0b6afc6
Update projects ignored for unreleased changes

The script for listing unreleased changes was ignoring oslo.version,
which has moved to the attic and is no longer part of Oslo. It was not
ignoring oslo-specs, which never has releases. Update the filter
patterns for the default list to always ignore those repositories.

Change-Id: I5e964da1964fa2898d56fe85015223e7c1c3662f