Updated openstack/openstack Project: openstack/cinder  a98aab6290a44108e3814fbf7880a5a275d4e512
Dell Storage Center API change fails init_conn

An attribute in the dell storage center REST API containing
the world wide name of the target device changed.  This meant
that on intialize connection we were returning None instead of
the wwn.

The driver has been changed to look for the new attribute and
to fall back to the old.  If neither is found it will not
return a target wwn of None.

Change-Id: I6ae6870d46937089bd6ed8799895f7068eebf01d
Closes-Bug: 1435935