Updated openstack/openstack Project: openstack/trove-integration  634e99b5cee521ba3bf60885a9f4d5150164ed3e
Fix the start / stop commands in redstack to work correctly

The current start / stop commands in redstack were using their
own mechanisms to send characters to the screen, and this is
causing a mismatch between what devstack considers the state
of the stack, and what the state of the stack actually is.

This commit updates the start and stop commands to use the
start and stop functions defined in the devstack lib to make
sure that consistency is achieved.

With this fix, developers should now be able to stop, start,
stop-deps, and start-deps consistently without any issues.

Change-Id: I8a692212a6a59125faeda52a7f588f73bb2c18ab