Updated openstack/openstack Project: openstack/python-novaclient  0343dff9735887ddcdf977c78f16fdb46a7ed09c
Removes reference to v3 nova api from novaclient docs

The Nova V3 API no longer exists. Although a version of the framework
exists in Nova as part of the new V2.1 API which is equivalent to
the V2 API there is no need to make mention of V3 in the novaclient
documentation.

In future we will be distinguish between legacy Nova API support (one
implementation of V2) and microversions Nova API support (the
new one) to reduce confusion around version numbers which will change quite
a bit without the framework changing.

Co-Authored-By: Andrey Kurilin <akurilin@mirantis.com>

Change-Id: Id0ff51e1165cb267045d7a63aff13c0e41336738