Updated openstack/openstack Project: openstack/cinder  7be2ac940f93cdda6a7b220b6b494fb525fafee2
Enable request-id in cinder API logs

Using common.local:store.context doesn't work correctly with
the new oslo_log module.  This change removes the calls to
common.local and removes the file altogether which allows the
proper settings to be passed in to the logging module and give
us back our request-id in the API logs.

Change-Id: If312f419fb465c9128bcfa0a280d6c35a46cb29b
Closes-Bug: #1435588