Updated openstack/openstack Project: openstack/neutron  ca620cc1368508ab47f997fa5e256ce2ba116532
Add eventlet monkey_patch helper

Eventlet monkey patching the os and thread modules causes subprocess.Popen
to fail on Windows when using pipes due to missing non-blocking IO support.
There is an reported bug on eventlet [1] with this issue. The solution
for this issue is to not monkey_patch os and thread.

This change is needed in order to run ovs_neutron_agent on Windows platforms.

[1] https://bitbucket.org/eventlet/eventlet/issue/132/eventletmonkey_patch-breaks

Partially implements blueprint: hyper-v-ovs-agent

Change-Id: I73f5dfc16563da2c09b5440a687484ba5a1fc2de