Updated openstack/openstack Project: openstack/tripleo-incubator  bb9670ce3d3c8383a49f4f62e02c44a92892b275
Set NeutronPublicInterface defaults to nic1

Updates the devtest_overcloud.sh and devtest_undercloud.sh
scripts so that we use the nic1 naming abstraction as a
default value. This matches what we already do in
the tripleo-heat-templates and is a better default in
that it likely works in both virt and most baremetal
environments.

Without this fix if you are deploying to real hardware eth0 may
not exist on some distributions (RHEL, Fedora) and by default
your installation would fail.

Change-Id: I4dc6152adf1d9faebcd1e8eaba37c563b03b7162