Updated openstack/openstack Project: openstack-infra/puppet-nodepool  a67f1013392eec9f73a39a7fd6f20abb059e1491
Install additional depends for dib

Building minimal centos, fedora and ubuntu images into which we put the
rackspace network handling code requires a few more packages than what
we install now.

Also, for now, the vhd-util code is in a PPA on Launchpad. We need a
better solution for this, but hopefully this will get us by for the time
being.

Depends-On: I6f15c5c068901120e1efff5988aabda370bf5df9
Change-Id: I9d49d88d38780c89e499a7a25e4c648d9e6bdf56