Updated openstack/openstack Project: openstack/tempest-lib  c20857ca1ef9b61bbb814a961144b61e4f33c108
Add optional prefix parameter to rand_name function

This commits an optional prefix kwarg to the rand_name function in
data utils. The intet of this argument is to allow a consumer of the
function to wrap randname to specify a project specific prefix to
all the generated names. For example, this will be useful for
identification of test created resources in case of bugs causing
resource leaks.

Change-Id: I7f7a1313ccbd8c5ef1e720537c16d60c89ed5923