Updated openstack/openstack Project: openstack/nova  42d020db5f6a5c3ebd0d06e397a8639d7876c2c0
Add a fixture for the NovaObject indirection API

Before the service code can be converted to objects, a fixture to
handle the global NovaObject indirection api is needed. Without it
some tests would fail with a messaging timeout due to requests for
a nonexistent conductor service because the global indirection_api
would mistakenly be set.

Related to blueprint kilo-objects

Change-Id: I2d417923c09df3a7923e5f15e2e8cf41b40f4f62