Updated openstack/openstack Project: openstack/neutron  88f7e3aed68e3abf6f9d97c26ad9ba6d01d34e7b
Fix intermittent failure in TestNetworksFailover UT

Ensure that the periodic check does not get in the way of method
calls being tested, by stopping the periodic task from running.

This patch moves the mock for the periodic check into the
setup_coreplugin call so it gets called by unit tests that
use the core plugin.

The previous location after the construction of the API router was
too late because the core plugin was already constructed by the
neutron manager. This led to random failures because the periodic
tasks leaked by all of the unit tests would occasionally preempt
test_reschedule_network_from_down_agent in automatically removing
a network from an agent.

Co-Author: Jenkins <jenkins@review.openstack.org>

Change-Id: I60ad7fa8ca874f93b7f806a0e035be84180a5de9
Closes-bug: #1432958