Updated openstack/openstack Project: openstack-infra/project-config  8c8d8a4d09e4568b662ace3915f89628a44450d7
Revert "Promote devstack centos7 to non-voting check job"

The centos7 jobs are regularly failing in a way that
triggers an OOM of the jenkins process. Which causes jobs
to restart and things to effectively get stuck in the check
queue. It's not an 100% occurance, however it's creating
substantial delays on selected patches.

We should put the centos7 job back to experimental until 
the issue is addressed.

This reverts commit 8ff84b99b3df57c7e362d7d08cfc75d17a520aca.

Change-Id: Ib2f2f898ac9e8dce0e65b50aa6f353bc1b049146