Updated openstack/openstack Project: openstack/taskflow  c85805fca1df9b796e8e4fd0e3b7b9248be7267a
Prefer posixpath to os.path

To ensure that the memory fake filesystem works uniformly
across distributions prefer to use the posixpath module
which always works with '/' and friends instead of the
os.path and os.path.sep and friends which may vary depending
on operating system used.

Since we have tested with the usage of '/' we might as well
just restrict to that; and avoid the cross operating system
issues that may pop up when using this fake filesystem.

In general isolating one self from the operating system specifics
is IMHO preferable for this; as it avoids edge cases that we
don't care about.

Change-Id: I3f61f380e1bcb131bc42b627adf9dfe8a7f2d992