Updated openstack/openstack Project: openstack/cinder  af809cdab14a6357162f179aa3591507978e5170
Added the missing attachment to detach_volume

I must have missed this driver in the multiattach
patch.   The attachment object is passed in to
detach_volume to give drivers a chance to deal with
a detach call.

This patch just updates the signature of the method
to be the same as the parent.

Change-Id: I895c163e67d35df202a5490a7a31ee7fc13a0cb4
Closes-Bug: #1436367