Updated openstack/openstack Project: openstack/keystone  28b6f45ac10a9ed4cbb155c95082a3945deafbf9
Loosen the validation schema used for trustee/trustor ids

Previously, the jsonschema validator would match identifier strings to a
regular expression representing something similar to a UUID. This is not always
the case depending on how the user may have identifier strings setup in their
deployment. This change allows for periods to be contained within an identifier
string.

Change-Id: I18d8a0347bab7c8c403368c53d24d144d36aa093
Closes-Bug: #1407661