Updated openstack/openstack Project: openstack/cinder  7627531262105c5a473e07184ba0e898d1da7e1c
Sort list of cinder_object.changes

If PYTHONHASHSEED is set to a non-zero value,
test_objects.TestObject.test_base_attributes unit test
fails.  This is because 'cinder_object.changes' key is
an unsorted list when the object is converted to a
primitive.  The following patch fixes this by sorting
cinder_object.changes.

Change-Id: I3a80396c5c897de26e982fa6426e7720a09c1375
Closes-Bug: #1434154