Updated openstack/openstack Project: openstack/cinder  2b19cfc2657c3eca363b0dba00600055c10585d5
Fixed access to instance_uuid in retype

We missed an access to volume['instance_uuid']
in the multiattach patch.   The instance_uuid and attached_host
were moved into the volume_attachment table.

This patch changes how we test to see if a volume is available
or not, but checking the new location.

Change-Id: I8844cd8c2947845f9efb11a2a1371efee2d28bbe
Closes-Bug: 1436604