Updated openstack/openstack Project: openstack/nova  a787d7604306975fa0e65f2d8e43679a15b3b4f6
Adds cleanup on v2.2 keypair api and tests

Removes unused arguments in test_keypairs.
Sets the keypair 'type' field to enum in the v2.2 validation
schema.
Removes check if the keypair 'type' is valid, as it will be done
by the validation schema.
Adds functional tests for ssh and x509 keypairs.

Closes-Bug: 1434033

Change-Id: I39f0a363458a17a4011a0661a45dc449f122ba1a