Updated openstack/openstack Project: openstack-dev/devstack  a22942ba52917dcdfa031f38301742884dfb9b4e
Recreate tox venv for tempest on stack.sh run

Sometimes when certain dependencies are changed, recreating an existing
DevStack fails, as tempest is using an old existing venv for tox run,
where dependencies are not met.

This change should help developers who are reusing
their existing DevStack.
The gate should not be affected as there the vev is freshly created
anyway.

Change-Id: Ic42ba1cb0aa829c5120151d3d8cdafa4efc3ffaa