Updated openstack/openstack Project: openstack/trove-integration  dc0c9f57becbe13207c536d258f71725758f3e5a
Add SERVICE_DIR to redstack configuration

Currently redstack allows overriding of DEST as the base directory
for OpenStack components.  The related subdirectory DATA_DIR is
then set to point underneath it, however SERVICE_DIR is not.  This
means that the pid files are still written into /opt/stack/status
instead of a directory under the new DEST location.

SERVICE_DIR is now set in localrc.rc and redstack.rc

Change-Id: I2b446e9d960bc9c7e54252953d78d3c93b0e1d58