Updated openstack/openstack Project: openstack/nova  8ef9d45c9aa103cb76c90658c798d52faefb6938
Fix missing the cpu_pinning request

When boot instance with dedicated cpu_policy, the cpu_pinning field
of InstanceNUMATopology is '{}'. But the function
'instance_topology_from_instance' just missed that fields when convert
instance numa topology dict back to object. That leads to the cpu_pinning
just ignored. This patch fixed that.

Change-Id: I8711cdc59c7455b3c8f94c68ab7f38d96ce168c4
Closes-Bug: #1436823