Updated openstack/openstack Project: openstack/nova  86c861119b5048e04b7a8b925fcce9f7c67620b5
Tidy up sentinel comparison in pop_instance_event

This is a non functional change, as object().__eq__ defaults to 'is'
in any case. The intended behaviour here is 'is', so this just makes
that explicit, and therefore slightly clearer. It might be a
femtosecond faster too due to less indirection.

Change-Id: I156b6463a126a2b1684fdea6ff120a5e28e6be69