Updated openstack/openstack Project: openstack-infra/system-config  fc951f3a2ee84c0401c08b61cbbc50e07787649c
Apply F20 gruby work-around only on RAX

The issue with F20 grubby killing the extlinux.conf file is specific
to RAX, because they ship a different file.  So apply this only when
on RAX because it is causing issues elsewhere.

This is not needed long-term.  We are close to removing f20 nodes
anyway with just a few jobs, all of which can be converted to centos.
All changes to do this are in-flight, see the dependency chain from
[1].  At that point all this goes away anyway.

[1] https://review.openstack.org/167443

Change-Id: Ibb5a455d690aa1298c2f625da20e606914bf632a