Updated openstack/openstack Project: openstack/keystone  6b41dd558f9c37cd83cf74dd13b4ad1e866eb19c
Refactor _create_projects_hierarchy in tests

The method _create_projects_hierarchy was
creating a project hierarchy having the project
which has id equals to self.project_id as root.
However, this project already contains role
assignments granted when loading sample data.

This patch refactors this method in order to
return a clean project hierarchy and re-use it
on existing methods that need a clean project
hierarchy.

Change-Id: If448d732a33a29eabf88d4ca88f676d26746ebea