Updated openstack/openstack Project: openstack-infra/project-config  233df7eefe68f00e2741b7a132e904dac3376f8f
Temp - making horizon integration tests non-voting

The horizon integration tests are failing consistently, not based
on the content of the patch. The particular integration test that
fails is not indicative of cause, as the currently offending test
can be skipped and a different integration test fails.

This needs more time to diagnose, but is currently blocking all
horizon check and gate jobs from passing.

Related-Bug: #1436903
Change-Id: I8ad2ead8c37b0cce73393188c05e5ba4ef5b826e