Updated openstack/openstack Project: openstack/nova  d0c2684ab8816d30d72a75f7b3e1440b49755f0f
libvirt: Resize down an instance booted from a volume

When we boot instance from volume, it doesn't use root disk and
we should provide the possibility to resize down this instance.
This patch allows to change the size down of the instance
booted from volume.

Also this patch got rid of duplicate code in test-cases.

Closes-Bug: #1416361
Change-Id: I0f890de69d268218d44a6ab4117a5c69cf6fbf8f