Updated openstack/openstack Project: openstack/neutron  2665b7592e751a69a8719414d1bda69b70a1fb94
If configured, set the MTU for fpr/rfp intefaces

if network_device_mtu parameter is configured, set the MTU for
fpr and rfp interfaces to the value specified by the parameter
at the time of creation of these interfaces.

Enhanced DVR functional test to verify MTU gets
set for the fpr/rfp interfaces.

Co-Authored-By: Adolfo Duarte <adolfo.duarte@hp.com>
Closes-bug: #1429162
Change-Id: Ie41122d1f7306dfd3debbbb8dbf2ecabf716dcb8