Updated openstack/openstack Project: openstack/neutron-vpnaas  1864d99eb145c8fcf0eebb1f4b48a6c97260fb85
VPNaaS Fix unit test breakage

Changing path to execute() so that it matches neutron changes.

Will do a separate commit to clean up the test module better so that
it doesn't rely on Neutron for this mock (I tried it earlier today,
but I found other issues and need more time for a robust fix).

Change-Id: Id9838233f75a43ae2fa68032251cd624abe38528